requirements.txt
langchain
openai
python-environ
transformers
accelerate
einops
xformers